pc105: Show one more key on the 4th row
A significant amount of keyboards have 105 keys (e.g. scandinaviad
keyboards), while the common US keyboard only have 104. The difference
is the extra key right of the left shift key. Show the 105th key to make
that key discoverable. It means there will be a key on US keyboards that
they can't find on the actual keyboard, but that's probably better than
not being able to find some key that is actually there.